forward_proxy/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
httpx
sqlalchemy
python-multipart
python-dotenv
requests
Pillow
openai
